requirements.txt
streamlit==latest
pillow